Modified the readCSV function to output a vector of data so later be used to build the max heap within the reccomend function. Implemented the quicksort as well as the quicksort reccomend function

+++ src/MaxHeap.h
@@ -66,7 +66,7 @@ public:
         return heap.empty();
     }
 
-    size_t size() const {
+    int size() const {
         return heap.size();
     }
 };
